requirements.txt
dj-config-url
django==2.2
django-ckeditor
django-material
Pillow
django-heroku
gunicorn
whitenoise
django-storages